Fix loading data to member profile

+++ android/engine/src/main/java/org/smartregister/fhircore/engine/data/local/patient/dao/register/DefaultPatientRegisterDao.kt
@@ -62,7 +62,7 @@ constructor(
           id = it.logicalId,
           name = it.extractName(),
           gender = it.gender,
-          age = it.extractAge().toInt()
+          age = it.extractAge()
         )
       }
     }
